clear warnings on mod_*

+++ modules/mod_comment.php
@@ -19,14 +19,14 @@
 	return $value;
 	}
 	 
-	if ($function == 'Add Game Comment') {
+	if (isset($function) && $function == 'Add Game Comment') {
 		$comment = htmlspecialchars($comment, ENT_QUOTES);
 		$userid = check_input($userid);
 		$gameid = check_input($gameid);
 		$commentQuery = mysql_query(" INSERT INTO comments (userid, gameid, comment, timestamp) VALUES ('$userid', '$gameid', '$comment', FROM_UNIXTIME($time)) ") or die('Query failed: ' . mysql_error());
 	}
 
-	if ($function == 'Delete Game Comment') {
+	if (isset($function) && $function == 'Delete Game Comment') {
 		$commentQuery = mysql_query(" DELETE FROM comments WHERE id = $commentid ") or die('Query failed: ' . mysql_error());
 	}
